extract CacheInspector
