+++ cache/CacheInspector.tsx
@@ -1,0 +1,3 @@
+import { Director } from 'hitchcock';
+
+export default Director;
