+++ cache/reactCache.tsx
@@ -6,6 +6,5 @@ export const cache = createCache();
 // hitchcock uses its own cache, so we ignore it here
 export const createResource = (args, ...hash) => {
     const resource = createResourceViaHitchcock(args, ...hash);
-    return {read: (_cache, args) => resource.read(args)};
+    return { read: (_cache, args) => resource.read(args) };
 }; // Provided by 'simple-cache-provider'
-
